+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+- Allow to provide the login URL through the recording panel.
+
 ## 0.0.14 - 2025-05-23
 
 ### Fixed